:recycle: Moved router to plugins
