+++ refarch-frontend/src/plugins/index.ts
@@ -1,8 +1,8 @@
 import type { App } from "vue";
 
 import pinia from "@/plugins/pinia";
+import router from "@/plugins/router";
 import vuetify from "@/plugins/vuetify";
-import router from "@/router";
 
 export function registerPlugins(app: App) {
   app.use(vuetify).use(router).use(pinia);
